(MODULES-3035) Manage config settings
Provide the ability to manage config settings with a custom type and
provider. This allows a config setting to be ensurable, and also
allows for a config settings to be manageable in the same converge
that Chocolatey is installed.

The minimum version of Chocolatey has been set at 0.9.10 as that
is where unsetting configuration values was implemented.

When it comes to config values that contain passwords, those are
not ensurable, but the insync? will check to see if it has a value
or not and respond appropriately.