(MODULES-3391) Fix failure in config_spec.rb
Prior to this commit a test in the config_spec was failing with:

Mocha::ExpectationError:
unexpected invocation: #<AnyInstance:Win32::Registry>.[]('Domain')
satisfied expectations:
- allowed any number of times, invoked 4 times:
  #<AnyInstance:Win32::Registry>.[]('ChocolateyInstall')
- allowed any number of times, not yet invoked:
  #<Puppet::Util::Feature:0x45b5be0>.root?(any_parameters)

This is because a facter fact was making a call to Win32::Registry.[] which
hadn't been stubbed (two, in fact). This commit adds these stubs to the
spec_helper in addition to the existing one. This commit makes test pass, but
may be more brittle than a more generic stub that doesn't specify the
parameter. This commit assumes the specificity is of higher value.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>