(MAINT) Fix CI download logic
For some reason when hitting the artifactory URL for the first time
the wrong package version is returned. This is true in the browser,
so it is not a problem in the code path.

This commit adds logic to the helper function to hit the URL twice
in order to force the correct version to be returned.